+++ src/bss_web_file_server/services/image.py
